+++ test/jest/helpers/index.js
@@ -2,3 +2,4 @@ export * from './renderWithReduxForm';
 export * from './renderWithFinalForm';
 export * from './stripesMock';
 export * from './translationsProperties';
+export * from './shortcuts';