+++ src/components/ProfileTree/ProfileLinker/ProfileLinker.js
@@ -4,7 +4,6 @@ import React, {
 } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { noop } from 'lodash';
 
 import classNames from 'classnames';
 
@@ -111,7 +110,7 @@ export const ProfileLinker = ({
           id={`type-selector-dropdown-${id}`}
           className={classNames(css['linker-button'], className)}
           open={typeSelectorOpen}
-          onToggle={noop}
+          onToggle={() => setTypeSelectorOpen(!typeSelectorOpen)}
           renderTrigger={trigger}
           renderMenu={menu}
           usePortal={false}