+++ src/settings/MappingProfiles/ViewMappingProfile.js
@@ -377,6 +377,7 @@ export class ViewMappingProfile extends Component {
                   parentType={PROFILE_TYPES.MAPPING_PROFILE}
                   masterType={PROFILE_TYPES.ACTION_PROFILE}
                   detailType={PROFILE_TYPES.MAPPING_PROFILE}
+                  profileType={ENTITY_KEYS.MAPPING_PROFILES}
                   profileName={name}
                   contentData={associations}
                   hasLoaded={hasLoaded}