+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Cover `<ListTemplate>` subcomponent with tests (UIDATIMP-708)
 * Change MARC SRS bib column header and toggle (UIDATIMP-1011)
 * Add hotlinks to the Import log summary for EDIFACT imports (UIDATIMP-989)
+* Cover `<DetailsKeyShortcutsWrapper>` component with tests (UIDATIMP-956)
 
 ### Bugs fixed:
 * Auxiliary "repeatableFieldAction" property disappear while removing "vendor reference number" field mapping (UIDATIMP-987)