+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Cover `<DetailsKeyShortcutsWrapper>` component with tests (UIDATIMP-956)
 * Get rid of outdated componentWillReceiveProps method in SearchAndSort component (UIDATIMP-1001)
 * Cover `<EditKeyShortcutsWrapper` component with tests (UIDATIMP-957)
+* refactor psets away from backend ".all" permissions (UIDATIMP-1017)
 
 ### Bugs fixed:
 * Auxiliary "repeatableFieldAction" property disappear while removing "vendor reference number" field mapping (UIDATIMP-987)